feat: lazy load request on collection popover
